Define _gid and gid in Node.prototype.
The getter of gid will assign a new _gid value if not already assigned. This makes sure that whenever we want to read the gid value of a node, the value won't be undefined.

By this, we also get rid of the ugly phase of assign initial GID in DOMContentLoaded.